sale_alert/requirements.txt
aiohttp==3.9.5
pyyaml==6.0.1
python-dotenv==1.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
aioresponses==0.7.6
